miner: address BidBlock review feedback
